added useChains hook
